@pangloss 2015-12-08
Hi David,

Good catch!

It definitely should not be getting the indices twice. At the same time, I don't think it would make any sense to produce a list of both vertex and edge indices. They probably wouldn't even be distinguishable, at least not reliably across graphs...

Probably makes more sense to just default the type argument of that method to :vertex.

As for memoization, that's probably best to make Titan-specific. The best way to do that is probably to override the method in the Titan adapter. Perhaps a TTL based cache with invalidation on local index operations?

Thanks,
Darrick