@dstuebe 2015-12-08
Okay - will modify the PR tonight as suggested and tag you on changes in pacer-titan as well.
Thanks
David