@dstuebe 2015-12-10
@pangloss Does this look about right to fix the bug for nil type?