@pangloss 2015-12-10
That looks good, thanks!